Rename util exec to shell. Fix failed tests of "getOptions" util.
