+++ lib/commands/bootstrap.js
@@ -10,7 +10,7 @@ const path = require( 'upath' );
 const buildOptions = require( 'minimist-options' );
 const minimist = require( 'minimist' );
 const chalk = require( 'chalk' );
-const exec = require( '../utils/exec' );
+const shell = require( '../utils/shell' );
 
 module.exports = {
 	get helpMessage() {
@@ -58,7 +58,7 @@ module.exports = {
 				`git checkout --quiet ${ data.repository.branch }`
 			].join( ' && ' );
 
-			promise = exec( command );
+			promise = shell( command );
 		}
 
 		return promise
