+++ lib/commands/exec.js
@@ -7,8 +7,8 @@
 
 const fs = require( 'fs' );
 const path = require( 'upath' );
-const exec = require( '../utils/exec' );
 const chalk = require( 'chalk' );
+const shell = require( '../utils/shell' );
 
 module.exports = {
 	get helpMessage() {
@@ -54,7 +54,7 @@ module.exports = {
 
 			process.chdir( newCwd );
 
-			exec( data.arguments[ 0 ] )
+			shell( data.arguments[ 0 ] )
 				.then( stdout => {
 					process.chdir( data.mgitOptions.cwd );
