+++ lib/utils/getoptions.js
@@ -20,7 +20,7 @@ module.exports = function cwdResolver( callOptions, cwd ) {
 	let options = {
 		cwd,
 		packages: 'packages',
-		resolverPath: path.resolve( __dirname, '../default-resolver.js' ),
+		resolverPath: path.resolve( __dirname, '..', 'default-resolver.js' ),
 		resolverUrlTemplate: 'git@github.com:${ path }.git',
 		resolverTargetDirectory: 'git',
 		resolverDefaultBranch: 'master',
