+++ tests/fixtures/project-with-options-in-mgitjson/mgit.json
@@ -1,6 +1,5 @@
 {
   "packages": "foo",
-  "recursive": true,
   "dependencies": {
     "simple-package": "a/b"
   }
